+++ CHANGELOG.md
@@ -12,8 +12,8 @@
 
 ### Enhancements
 
-* Add `multiple_statements_declaration` opt-in rule 
-  that triggers when the there's multiple statements at the same line.  
+* Add `multiple_statements` opt-in rule
+  that triggers when there are multiple statements on the same line.
   [Ahmad Zaghloul](https://github.com/AhmedZaghloul19)
 
 * Exclude types with a `@Suite` attribute and functions annotated with `@Test` from `no_magic_numbers` rule.